lookahead: add new extended map lookahead
This lookahead generates a map with a more complex and complete way of
sampling and expanding the nodes to reach all destinations.

The dijkstra expansion is parallelized to overcome the higher number of
samples needed.

It also adds a penalty cost factor based on distance between the origin
and destination nodes.

Signed-off-by: Alessandro Comodi <acomodi@antmicro.com>